refa: get rid of home route
